refactor: move files around more logically
