+++ src/cddl/src/cddl/parser.rs
@@ -1,0 +1,3 @@
+#[derive(Parser)]
+#[grammar = "cddl/cddl.pest"]
+pub struct CDDLParser;
